### 2020-06-17 #2300 @Kevin-Mattheus-Moerman
@skapfer thanks for your submission. I will be looking for editors to handle this submission. In the mean time I recommend that you based on our [guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) you make some changes to your paper. In particular can you work on:

> - A clear Statement of Need that illustrates the research purpose of the software.
> - A list of key references, including to other software addressing related needs.

In relation to the `statement of need`, can you more clearly describe the research applications for this work? Currently your first sentence hints at the importance of these methods to science but a dedicated short paragraph with clear examples would be a great addition.
In relation to `key references`, can you refer to some existing codes that perform a similar task? Can you compare/constrast them with your library briefly?
