### 2020-06-28 68a509c
Extend introduction, and add literature
