+++ paper/paper.bib
@@ -1,18 +1,6 @@
 @STRING{AAP = {Astronomy \& Astrophysics}}
 
 
-@ARTICLE{bib:AdvMatReview,
-  doi = {10.1002/adma.201100562},
-  title = {Minkowski Tensor Shape Analysis of Cellular, Granular and Porous Structures},
-  url = {http://dx.doi.org/10.1002/adma.201100562},
-  journal = {Advanced Materials},
-  author = {G.E.~Schr\"oder-Turk and W.~Mickel and S.C.~Kapfer and M.A.~Klatt and F.M.~Schaller and M.J.F.~Hoffmann and N.~Kleppmann and P.~Armstrong and A.~Inayat and D.~Hug and M.~Reichelsdorfer and W.~Peukert and W.~Schwieger and K.~Mecke},
-  number = {22-23},
-  volume = {23},
-  year = {2011},
-  pages = {2535--2553}
-}
-
 @article{bib:Mantz2008,
   author={Hubert Mantz and Karin Jacobs and Klaus Mecke},
   title={Utilizing {Minkowski} functionals for image analysis: a marching square algorithm},
@@ -32,8 +20,8 @@
     volume = {2010},
     number = {11},
     pages = {P11010},
-    author = {Kapfer, S C and W Mickel and Schaller, F M and M Spanner and C Goll and T Nogawa and N Ito and K Mecke and Schröder-Turk, G E},
-    title = {Local anisotropy of fluids using Minkowski tensors},
+    author = {Sebastian C. Kapfer and W Mickel and Schaller, F M and M Spanner and C Goll and T Nogawa and N Ito and K Mecke and Schröder-Turk, G E},
+    title = {Local anisotropy of fluids using {Minkowski} tensors},
     journal = {Journal of Statistical Mechanics: Theory and Experiment},
 }
 
@@ -42,22 +30,148 @@
   title = {Shortcomings of the bond orientational order parameters for the analysis of disordered particulate matter},
   url = {http://link.aip.org/link/?JCP/138/044501/1},
   journal = {The Journal of Chemical Physics},
-  author = {Walter Mickel and Kapfer, Sebastian C. and Schr\"oder-Turk, Gerd E. and Klaus Mecke},
+  author = {Walter Mickel and Sebastian C. Kapfer and Gerd E. Schr\"oder-Turk and Klaus Mecke},
   number = {4},
   volume = {138},
   year = {2013},
   pages = {044501:1--7}
 }
 
+@Inbook{bib:Springer2017,
+  author={Michael A. Klatt and G{\"u}nter Last and Klaus Mecke and Redenbach, Claudia and Fabian M. Schaller and Gerd E. Schr{\"o}der-Turk},
+  editor={Jensen, Eva B. Vedel and Kiderlen, Markus},
+  title={Cell Shape Analysis of Random Tessellations Based on {Minkowski} Tensors},
+  bookTitle={Tensor Valuations and Their Applications in Stochastic Geometry and Imaging},
+  year={2017},
+  publisher={Springer International Publishing},
+  pages={385--421},
+  isbn={978-3-319-51951-7},
+  doi={10.1007/978-3-319-51951-7_13},
+  url={https://doi.org/10.1007/978-3-319-51951-7_13}
+}
+
+
+@ARTICLE{bib:AdvMatReview,
+  doi = {10.1002/adma.201100562},
+  title = {Minkowski Tensor Shape Analysis of Cellular, Granular and Porous Structures},
+  url = {http://dx.doi.org/10.1002/adma.201100562},
+  journal = {Advanced Materials},
[…]
