### 2020-06-30 #2300 @skapfer
@hugoledoux Sorry for the delay. We extended the introduction and added a few more references to applications of Minkowski Tensors. Here's some more ideas for referees:
spolavar
ebertin
rasband
j0r1
pmelchior

Best, Sebastian (on behalf of all authors)
