### 2020-07-31 #2300 @hugoledoux
👋 @skapfer I am having *real* problems finding a second reviewer, the ones you proposed all declined 🙈  (because no C++, or no time). So if you have more than please help, I am waiting for answers from a suggestion from one (but it's holidays...).

I'll start the official review with one, and add another when the person is found.

I'm away on holidays myself the next 2 weeks also... 🏕️
