### 2020-07-31 #2538 @whedon
[REVIEW]: papaya2 & Morphometer: 2D Irreducible Minkowski Tensor computation
**Submitting author:** @skapfer (<a href="http://orcid.org/0000-0002-7591-2739">Sebastian Kapfer</a>)
**Repository:** <a href="https://github.com/morphometry/papaya2" target ="_blank">https://github.com/morphometry/papaya2</a>
**Version:** v1.0.0
