### 2020-08-15 #2538 @kenohori
Hello,

Started working on the review today and noticed a few things. Regarding the code:
* The Makefile broadly works, but it doesn't look for paths, and so is not really an automated solution. Include and library paths need to be added manually. It doesn't link (-l) against python (for pypaya2).
* The Python module is 2.x only, which was retired on January 1.
* The software fails to build with CGAL 5.x (because it needs -std=c++14). IMO, it should also be updated to use the newer header-only mode of CGAL.

Regarding the documentation:
* There is no README information. The link to the website works, but this is not really a good practice (not versioned, no instructions when checking out repo, can be removed anytime).
* There is no clear list of dependencies. Some are listed throughout the documentation, but this is not ideal. numpy is not mentioned.
* Doesn't include information about test data or testing in general.
* There are no instructions for third parties to contribute to the software.

Finally, I want to know what is the link to Morphometer? Is Morphometer open source? Is it a web application only or is there some backend?
