### 2020-08-29 02e6a45
adjust paths and Makefile for tests
