+++ javascript/README.md
@@ -1,0 +1,7 @@
+# Usage
+
+Install [emscripten](https://github.com/emscripten-core/emscripten).
+(jspaya is tested with emscripten Version 2.0.1)
+
+To compile type `make`.
+
