### 2020-08-29 cfc4686
updated jspaya2 Makefile
