+++ python/README.md
@@ -1,0 +1,9 @@
+# Usage
+
+Install the header files for Python 3 and numpy.
+(On Debian-based systems: `apt get install python3-dev python3-numpy`.)
+
+To compile type `make`.
+
+You can put the resulting `pypaya2.so` in the same directory as your analysis
+Python scripts, or on your PYTHONPATH.
