### 2020-08-29 820bcb0
we have included the JS bindings; added READMEs
