+++ matlab/README.md
@@ -1,0 +1,5 @@
+# Matlab examples
+
+See `imganalysis.m` for an example script that builds and uses the image analysis Matlab extension.
+
+See `ppanalysis.m` for an example script that builds and uses the point pattern analysis Matlab extension.
