### 2020-08-29 #2538 @skapfer
Dear @kenohori ,
thank you for the constructive feedback. We have addressed your comments in the code repository and the paper.

Specifically:
- we have made sure that the software builds correctly with CGAL 5 and Python 3
- we have included the JavaScript binding in the repository
- we have reorganized the repository to more clearly identify the components
- we have added a README, including instructions on how to contribute and where to file bugs, and how to run the test suite
- we have added READMEs which detail installation requirements and instructions for the language bindings

The morphometer is a pure web application (no backend) which uses a version of papaya2 compiled to JavaScript. For completeness, we have added this binding to the code repository as well. The rest of the Morphometer is merely wrapping the library in a nice webpage. Its source code is GPLv3 (see imprint).

We hope that we could address your concerns and would like to thank you again for your feedback.

Sebastian & Fabian, on behalf of the authors
