### 2020-08-30 bd14c69
sk - a little less symmetric polygon
