### 2020-08-30 c473e03
sk - add test coverage
