### 2020-08-30 dde1583
sk - add imt_for_polygon to pypaya
