### 2020-08-30 e262c97
sk - add test coverage
