+++ python/pypaya2.cpp
@@ -1,7 +1,10 @@
 // Python/Numpy interface.
 // we're not using Boost::Python here as we export very few
 // functions and Boost::Python::Numpy is not in Debian.
-// 2019 Sebastian Kapfer <sebastian.kapfer@fau.de>
+// 2019-2020 Sebastian Kapfer <sebastian.kapfer@fau.de>
+// TODO: Extract common code between the two functions
+// TODO: Add test coverage for the pointpattern analysis
+
 #include <papaya2.hpp>
 #ifdef HAVE_CGAL
 // without CGAL, there is no Voronoi.
