### 2020-09-08 #2538 @rmjarvis
Some comments re installation:
* I had to manually add -std=c++11 (or later).  I don't think this breaks newer compilers and is required for older ones, so I'd recommend just adding that to the Makefile.
* I hit an error compiling the demos because I didn't have CGAL installed.  Nothing in the README says that this is required, so I'd recommend at least adding that instruction to the README file.  A better solution might be to explicitly describe the different demos you can make with their corresponding make commands (`make sersic`, `make ppanalysis`, etc.) and just point out that CGAL is required for `ppanalysis` and CCfits is required for banana.
* After installing CGAL (via conda install cgal), the compilation of ppanalysis failed:
```
$ make ppanalysis
c++ -g -O3 -std=c++11 -Wall -Werror -I../include -DHAVE_CGAL -c -o ppanalysis.o ppanalysis.cpp
In file included from /anaconda3/envs/py3.7/include/boost/random/detail/integer_log2.hpp:19:0,
                 from /anaconda3/envs/py3.7/include/boost/random/detail/large_arithmetic.hpp:19,
                 from /anaconda3/envs/py3.7/include/boost/random/detail/const_mod.hpp:23,
                 from /anaconda3/envs/py3.7/include/boost/random/linear_congruential.hpp:30,
                 from /anaconda3/envs/py3.7/include/boost/random/additive_combine.hpp:27,
                 from /anaconda3/envs/py3.7/include/boost/random.hpp:36,
                 from /anaconda3/envs/py3.7/include/CGAL/algorithm.h:39,
                 from /anaconda3/envs/py3.7/include/CGAL/Hilbert_sort_base.h:26,
                 from /anaconda3/envs/py3.7/include/CGAL/Hilbert_sort_median_2.h:27,
                 from /anaconda3/envs/py3.7/include/CGAL/Hilbert_sort_2.h:25,
                 from /anaconda3/envs/py3.7/include/CGAL/hilbert_sort.h:28,
                 from /anaconda3/envs/py3.7/include/CGAL/spatial_sort.h:26,
                 from /anaconda3/envs/py3.7/include/CGAL/Triangulation_2.h:45,
                 from /anaconda3/envs/py3.7/include/CGAL/Delaunay_triangulation_2.h:28,
                 from ../include/papaya2/voronoi.hpp:11,
                 from ppanalysis.cpp:11:
/anaconda3/envs/py3.7/include/boost/pending/integer_log2.hpp:7:89: note: #pragma message: This header is deprecated. Use <boost/integer/integer_log2.hpp> instead.
 BOOST_HEADER_DEPRECATED("<boost/integer/integer_log2.hpp>");
                                                                                         ^
c++ -g -O3 -std=c++11 -Wall -Werror -I../include -o ppanalysis ppanalysis.o  -lCGAL -LCGAL_Core -lgmp
ld: warning: directory not found for option '-LCGAL_Core'
$ ppanalysis
dyld: Library not loaded: @rpath/libCGAL.13.dylib
  Referenced from: /Users/Mike/papaya2/demos/./ppanalysis
  Reason: image not found
Abort trap: 6
```
* Installation of banana also failed:
```
$ make banana
c++ -g -O3 -std=c++11 -Wall -Werror -I../include -o banana banana.cpp  -lCCfits -lcfitsio
Undefined symbols for architecture x86_64:
  "CCfits::FitsException::FitsException(std::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, bool&)", referenced from:
      CCfits::Image<double>::readImage(fitsfile*, long, long, double*, std::vector<long, std::allocator<long> > const&, bool&) in ccnYZ9YS.o
      CCfits::Image<float>::readImage(fitsfile*, long, long, float*, std::vector<long, std::allocator<long> > const&, bool&) in ccnYZ9YS.o
      CCfits::Image<unsigned char>::readImage(fitsfile*, long, long, unsigned char*, std::vector<long, std::allocator<long> > const&, bool&) in ccnYZ9YS.o
      CCfits::Image<unsigned int>::readImage(fitsfile*, long, long, unsigned int*, std::vector<long, std::allocator<long> > const&, bool&) in ccnYZ9YS.o
      CCfits::Image<int>::readImage(fitsfile*, long, long, int*, std::vector<long, std::allocator<long> > const&, bool&) in ccnYZ9YS.o
      CCfits::Image<long long>::readImage(fitsfile*, long, long, long long*, std::vector<long, std::allocator<long> > const&, bool&) in ccnYZ9YS.o
      CCfits::PrimaryHDU<unsigned short>::readImage(long, long, unsigned short*) in ccnYZ9YS.o
      ...
  "CCfits::FITS::FITS(std::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, CCfits::RWmode, bool, std::vector<std::basic_string<char, std::char_traits<char>, std::allocator<char> >, std::allocator<std::basic_string<char, std::char_traits<char>, std::allocator<char> > > > const&)", referenced from:
      FitsFile::FitsFile(std::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, unsigned int) in ccnYZ9YS.o
  "CCfits::FITSUtil::UnrecognizedType::UnrecognizedType(std::basic_string<char, std::char_traits<char>, std::allocator<char> >, bool)", referenced from:
      CCfits::FITSUtil::MatchType<long long>::operator()() in ccnYZ9YS.o
  "CCfits::FitsFatal::FitsFatal(std::basic_string<char, std::char_traits<char>, std::allocator<char> > const&)", referenced from:
      void CCfits::PHDU::read<double>(std::valarray<double>&, long, long, double*) in ccnYZ9YS.o
ld: symbol(s) not found for architecture x86_64
collect2: error: ld returned 1 exit status
make: *** [banana] Error 1
```
I have CCfits installed (also via conda), and compiling with `c++ -v` shows that it is in my library search path, and the compiler didn't give an error about not finding libCCfits.dylib, so I think it found it.  But it seems to be trying to use a function that it can't find the link to.
* The unit tests fail with both g++ and clang.  First the g++ error: (c++ aliases to g++ on my system)
```
$ make
c++ -g -O3 -std=c++11 -Wall -Werror -I../include   -c -o test_image.o test_image.cpp
c++ -g -O3 -std=c++11 -Wall -Werror -I../include   -c -o test_tools.o test_tools.cpp
c++ -g -O3 -std=c++11 -Wall -Werror -I../include   -c -o test_imt.o test_imt.cpp
c++ -g -O3 -std=c++11 -Wall -Werror -I../include   -c -o runtests.o runtests.cpp
c++ -g -O3 -std=c++11 -Wall -Werror -I../include -o runtests test_image.o test_tools.o test_imt.o  runtests.o
./runtests
dyld: lazy symbol binding failed: Symbol not found: __ZNSt6chrono3_V212system_clock3nowEv
  Referenced from: /Users/Mike/papaya2/test/./runtests
  Expected in: /usr/lib/libstdc++.6.0.9.dylib

dyld: Symbol not found: __ZNSt6chrono3_V212system_clock3nowEv
  Referenced from: /Users/Mike/papaya2/test/./runtests
  Expected in: /usr/lib/libstdc++.6.0.9.dylib

make: *** [default] Abort trap: 6
```
* The error with clang is on compiliation:
```
$ CXX=clang++ make
clang++ -g -O3 -std=c++11 -Wall -Werror -I../include   -c -o test_image.o test_image.cpp
clang++ -g -O3 -std=c++11 -Wall -Werror -I../include   -c -o test_tools.o test_tools.cpp
test_tools.cpp:66:11: error: no matching member function for call to 'assign'
        v.assign(r.begin(), r.end());
        ~~^~~~~~
/anaconda3/envs/py3.7/bin/../include/c++/v1/vector:611:10: note: candidate function not viable: no
      known conversion from 'papaya2::IntegerRange::iterator' to 'std::__1::vector<int,
      std::__1::allocator<int> >::size_type' (aka 'unsigned long') for 1st argument
    void assign(size_type __n, const_reference __u);
         ^
/anaconda3/envs/py3.7/bin/../include/c++/v1/vector:599:9: note: candidate template ignored:
      requirement '!__is_cpp17_forward_iterator<papaya2::IntegerRange::iterator>::value' was not
      satisfied [with _InputIterator = papaya2::IntegerRange::iterator]
        assign(_InputIterator __first, _InputIterator __last);
        ^
/anaconda3/envs/py3.7/bin/../include/c++/v1/vector:609:9: note: candidate template ignored:
      requirement 'is_constructible<int, void>::value' was not satisfied [with _ForwardIterator =
      papaya2::IntegerRange::iterator]
        assign(_ForwardIterator __first, _ForwardIterator __last);
        ^
/anaconda3/envs/py3.7/bin/../include/c++/v1/vector:615:10: note: candidate function not viable:
      requires single argument '__il', but 2 arguments were provided
    void assign(initializer_list<value_type> __il)
         ^
1 error generated.
make: *** [test_tools.o] Error 1
```

Now I recognize that portability is hard, and I'm guessing none of the developers are using Macs.  These errors are all on a MacOS 10.13 system.  I'd recommend using a CI such as Travis to regularly build and run the tests and demos on a variety of systems including both Mac and Linux and ideally with at least a couple compilers.  This will help you determine what instructions to give about what is really required to make it work on different systems.  Especially with respect to dependencies, since Travis will make you install all of them explicitly.
