### 2020-09-08 #2538 @rmjarvis
Onto a third system now: SUSE Linux Enterprise Server 15, which has gcc 8.3.0.
* The unit tests compiled and ran successfully.
* I had to add the right `-I` and `-L` locations manually to tell it where CCfits and CGAL were installed.  But then `banana`, `sersic` and `imganalysis` all built successfully.
* `ppanalysis` compiled successfully, but failed the linking step:
```
$ make ppanalysis
g++ -g -O3 -Wall -Werror -I../include -I/global/u2/m/mjarvis/.conda/envs/my_root/include -L/global/u2/m/mjarvis/.conda/envs/my_root/lib -DHAVE_CGAL -c -o ppanalysis.o ppanalysis.cpp
g++ -g -O3 -Wall -Werror -I../include -I/global/u2/m/mjarvis/.conda/envs/my_root/include -L/global/u2/m/mjarvis/.conda/envs/my_root/lib -o ppanalysis ppanalysis.o  -lCGAL -LCGAL_Core -lgmp
/usr/bin/ld: cannot find -lCGAL
collect2: error: ld returned 1 exit status
make: *** [Makefile:48: ppanalysis] Error 1
```
I see a number of libraries that start with CGAL, but not one that is simple libCGAL.so:
```
$ ls /global/u2/m/mjarvis/.conda/envs/my_root/lib/libCGAL*
/global/u2/m/mjarvis/.conda/envs/my_root/lib/libCGAL_AABB_tree_cpp.so*
/global/u2/m/mjarvis/.conda/envs/my_root/lib/libCGAL_Alpha_shape_2_cpp.so*
/global/u2/m/mjarvis/.conda/envs/my_root/lib/libCGAL_Kernel_cpp.so*
/global/u2/m/mjarvis/.conda/envs/my_root/lib/libCGAL_Mesh_3_cpp.so*
/global/u2/m/mjarvis/.conda/envs/my_root/lib/libCGAL_Surface_mesher_cpp.so*
/global/u2/m/mjarvis/.conda/envs/my_root/lib/libCGAL_Triangulation_2_cpp.so*
/global/u2/m/mjarvis/.conda/envs/my_root/lib/libCGAL_Triangulation_3_cpp.so*
```
I don't know whether this is an error in the linking step of your code or in the conda installation of CGAL.  It seems plausible that there should be a single library to link to that encompasses all of these, which would logically be called libCGAL.so.  But it doesn't seem to be installed along with the rest of these more specific libraries.
