### 2020-09-08 morphometry/papaya2#14 @skapfer
Test suite does not work on MacOS 10.13 (very recent clang)
$ CXX=clang++ make
clang++ -g -O3 -std=c++11 -Wall -Werror -I../include   -c -o test_image.o test_image.cpp
clang++ -g -O3 -std=c++11 -Wall -Werror -I../include   -c -o test_tools.o test_tools.cpp
test_tools.cpp:66:11: error: no matching member function for call to 'assign'
        v.assign(r.begin(), r.end());
        ~~^~~~~~
/anaconda3/envs/py3.7/bin/../include/c++/v1/vector:611:10: note: candidate function not viable: no
      known conversion from 'papaya2::IntegerRange::iterator' to 'std::__1::vector<int,
      std::__1::allocator<int> >::size_type' (aka 'unsigned long') for 1st argument
    void assign(size_type __n, const_reference __u);
         ^
/anaconda3/envs/py3.7/bin/../include/c++/v1/vector:599:9: note: candidate template ignored:
      requirement '!__is_cpp17_forward_iterator<papaya2::IntegerRange::iterator>::value' was not
      satisfied [with _InputIterator = papaya2::IntegerRange::iterator]
        assign(_InputIterator __first, _InputIterator __last);
        ^
/anaconda3/envs/py3.7/bin/../include/c++/v1/vector:609:9: note: candidate template ignored:
      requirement 'is_constructible<int, void>::value' was not satisfied [with _ForwardIterator =
      papaya2::IntegerRange::iterator]
        assign(_ForwardIterator __first, _ForwardIterator __last);
        ^
/anaconda3/envs/py3.7/bin/../include/c++/v1/vector:615:10: note: candidate function not viable:
      requires single argument '__il', but 2 arguments were provided
    void assign(initializer_list<value_type> __il)
         ^
1 error generated.
make: *** [test_tools.o] Error 1
