### 2020-09-08 #2538 @rmjarvis
Some high-level comments, mostly about the paper and documentation:
* You piqued my interest about the utility of this package with some statements about how it provides a richer description of the shape or geometry of objects than say the 2d inertia.  I work in gravitational lensing, so we regularly use 2d moments to describe the shape (and indirectly the shear) of galaxies.  It works well for weak lensing, but it's not great for medium to strong lensing.  So I was most attracted to the example you have of a strong lensing system.  However, I couldn't really figure out how I would use it to either detect or characterize strong lensing images.  The example has 4 numbers that get computed as a function of a threshold: area, perimeter, area fraction, and q_2.  The first three don't seem particularly useful, although possibly some combination like perimeter^2/area might be diagnostic.  q_2 seemed potentially interesting, but I couldn't find much discussion about it.  It sounds a bit like ellipticity (responding to the quadrupole moment).  Is it better in some way than standard ellipticities?  What intuition should I develop about it?  Or for the higher order Minkowski structure metrics?  I feel like the paper would be significantly helped by an example like this (not necessarily this one, but probably something connected to one of your demo examples would be fine) a bit more to show how it could be useful to a real science case.
* The paper includes the example of an anisotropic Gaussian random field from Morphometer, but again, there are just a handful of numbers produced for it without much explanation of what they might be good for.  So it's hard for me as someone not already steeped in the use of these statistics to know why I should care.  If you at least sketched out an overview of what kind of science you can do with this (say with maps of large scale structure or some material science data maybe), it would help make it a bit more concrete and potentially give me more of a clue why I might want to use it for something I'm working on.
* The paper has a link to this page: https://morphometry.org/theory/anisotropy-analysis-by-imt/.  This page feels to me like information that should be in the paper.  This was where I learned the most about what the numbers calculated by the software were really about.  It's fine for it to live on the web in addition to the paper, but I don't think a link to this page from the paper is sufficient.
* It's probably too late at this point, but it wasn't obvious to me why this rewrite of papaya deserved to be a brand new package rather than just papaya version 2.0.  Refactoring of code happens all the time, and it usually doesn't merit making a new package with a new name.  This is normally done with release versions, even when the code has been completely redone.
* I think the demos would benefit from including some example input files.  Perhaps the same files that are used in the online demo.  Then I could try them out immediately, look at the input and output files, mess around with some of the parameters to see how things change.  Having to supply my own input (e.g. FITS image) files is a pretty big barrier to this being a useful demo.  If I have a problem, I don't know whether the issue is with my input file or how I'm running it or what.  But with a sample input file, you can give explicit instructions on how you can run it and what you should see as output.
