### 2020-09-08 #2538 @rmjarvis
> The JOSS guidelines to authors suggest a length of 250-1000 words for the paper, and we're rather at the upper end of the scale already.

Ah, I didn't realize this.  I'm used to reading significantly longer papers, so this paper felt very short to me.  But if that's the case, probably my suggestions about increasing the scope of the paper are off base.  Maybe just some quick overviews of some of the ways you can use the code for science then, with pointers to the relevant papers that use these techniques for people to read up on further.  I think that would be helpful to motivate interest in the various ways one can use the software.
