### 2020-09-08 5c669e2
add granular-cryst-cluster.txt
