### 2020-09-09 #2538 @hugoledoux
Indeed, we try to keep the paper <1000words. The software and the docs are the main results for JOSS.
