+++ .github/workflows/linux-test.yml
@@ -1,0 +1,30 @@
+name: Papaya2
+
+on: [push]
+
+jobs:
+  linux-test:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Install packages
+      run: sudo apt-get install clang-7 g++-4.8 clang-10 python3-dev python3-numpy
+    - name: Compile and run test suite with default compiler
+      run: test/with c++
+    - name: Compile python3 module without CGAL
+      run: cd python; make && python3 test.py && make clean
+    - name: Compile and run test suite with clang 10
+      run: test/with clang++-10
+    - name: Compile and run test suite with clang 7
+      run: test/with clang++-7
+    - name: Compile and run test suite with GCC 4.8
+      run: test/with g++-4.8
+    - name: Compile demos without libraries
+      run: cd demos; make -j imganalysis sersic && make clean
+    - name: Install CGAL and FITS
+      run: sudo apt-get install libcgal-dev libccfits-dev
+    - name: Compile demos with libraries
+      run: cd demos; make -j imganalysis sersic ppanalysis banana && make clean
+    - name: Compile python3 module with CGAL
+      run: cd python; make && python3 test.py && make clean
