### 2020-09-11 d9d95c7
autodetect CGAL version
