+++ config/test-recent-cgal.cpp
@@ -1,0 +1,1 @@
+#include <CGAL/Compute_cone_boundaries_2.h>
