+++ config/test-any-cgal.cpp
@@ -1,0 +1,1 @@
+#include <CGAL/Delaunay_triangulation_2.h>
