+++ demos/Makefile
@@ -4,8 +4,6 @@ CXXFLAGS += -Wall -Werror
 FITS_LDFLAGS = -lCCfits -lcfitsio
 CXXFLAGS += -I../include
 
-include ../cgal.mk
-
 # modify defaults here if necessary
 -include features.mk
 
@@ -18,6 +16,7 @@ PRODUCTS = \
 default: .ts.mk.hpp imganalysis ppanalysis
 
 include ../config/shared.mk
+include ../config/detect-cgal.mk
 
 all: default banana sersic
 
@@ -30,9 +29,6 @@ banana: banana.cpp
 
 ppanalysis: ppanalysis.o
 	$(CXX) $(CXXFLAGS) -o $@ $< $(LDFLAGS) $(CGAL_LDFLAGS)
-
-ppanalysis.o: ppanalysis.cpp
-	$(CXX) $(CXXFLAGS) $(CGAL_CXXFLAGS) -c -o $@ $<
 
 sersic: sersic.o
 	$(CXX) $(CXXFLAGS) -o $@ $< $(LDFLAGS)
