### 2020-09-11 da788ad
rationalize headers and fix ts.hpp
