+++ demos/common.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <papaya2.hpp>
+#include ".ts.mk.hpp"
+#include "readarg.hpp"
