### 2020-09-11 fdd78b0
extensible clean target, factor out common code
