+++ test/Makefile
@@ -11,27 +11,15 @@ TEST_OBJECTS = \
     test_tools.o \
     test_imt.o \
 
-BINARIES = \
+PRODUCTS = \
     runtests \
 
 default: all
 	./runtests
 
+include ../config/shared.mk
+
 all: .ts.mk.hpp runtests
-
-# hack to make any code recompile if Makefile changes
-.ts.mk.hpp: $(MAKEFILES)
-	@touch $@
-
-# automagic dependencies
-depend: *.cpp *.hpp .ts.mk.hpp
-	$(CXX) $(CXXFLAGS) -MM *.cpp >.depend
--include .depend
 
 runtests: $(TEST_OBJECTS) runtests.o
 	$(CXX) $(CXXFLAGS) -o $@ $(TEST_OBJECTS) runtests.o $(LDFLAGS)
-
-clean:
-	rm -f $(BINARIES) *.o .ts.mk.hpp
-
-.PHONY: default all clean depend
