### 2020-09-11 6871eba
do not try to create timestamp file, include it in git
This is an implicit dependency of all targets, so not having it
breaks the build
