### 2020-09-11 9921244
note about common compilation problems
