### 2020-09-11 2852b89
exchanged point pattern example
