+++ config/test-recent-cgal.cpp
@@ -1,1 +1,1 @@
-#include <CGAL/Compute_cone_boundaries_2.h>
+#include <CGAL/Polygonal_surface_reconstruction.h>
