### 2020-09-11 1f97eaa
add missing iterator members
This broke compilation on recent clang
