### 2020-09-12 e0b4ca4
workaround for MacOS/clang misbehaving
