### 2020-09-12 dc9015c
fix linking of Python module on MacOS
