### 2020-09-12 edfbdae
add point pattern test
