+++ .github/workflows/linux-test.yml
@@ -13,7 +13,10 @@ jobs:
     - name: Compile and run test suite with default compiler
       run: test/with c++
     - name: Compile python3 module without CGAL
-      run: cd python; make && python3 test.py && make clean
+      run: |
+        cd python
+        make
+        make non_cgal_test
     - name: Compile and run test suite with clang 10
       run: test/with clang++-10
     - name: Compile and run test suite with clang 7
@@ -27,4 +30,8 @@ jobs:
     - name: Compile demos with libraries
       run: cd demos; make -j imganalysis sersic ppanalysis banana && make clean
     - name: Compile python3 module with CGAL
-      run: cd python; make && python3 test.py && make clean
+      run: |
+        cd python
+        make clean
+        make
+        make test
