+++ .github/workflows/mac-test.yml
@@ -30,14 +30,20 @@ jobs:
         make non_cgal_test
     - name: Compile demos without libraries
       run: cd demos; make -j imganalysis sersic && make clean
-    - name: Install CGAL
-      run: curl -L https://github.com/CGAL/cgal/releases/download/v5.1/CGAL-5.1.tar.xz | tar xJ
+    - name: Download Boost and CGAL
+      run: |
+        curl -L https://dl.bintray.com/boostorg/release/1.74.0/source/boost_1_74_0.tar.bz2 | tar xj
+        curl -L https://github.com/CGAL/cgal/releases/download/v5.1/CGAL-5.1.tar.xz | tar xJ
     - name: Compile demos with CGAL
-      run: cd demos; echo CXXFLAGS+=-I../CGAL-5.1/include >features.mk; make -j imganalysis sersic ppanalysis && make clean
+      run: |
+        cd demos
+        echo CXXFLAGS += -I../CGAL-5.1/include -I../boost_1_74_0 >features.mk
+        make -j imganalysis sersic ppanalysis
+        make clean
     - name: Compile python3 module with CGAL
       run: |
         cd python
         make clean
-        echo CXXFLAGS+=-I../CGAL-5.1/include >features.mk
+        cp ../demos/features.mk .
         make
         make test
