### 2020-09-14 590e59b
extend python readme
