### 2020-09-16 299848a
extend docs on how to compile the mexes
