+++ demos/README.md
@@ -13,7 +13,7 @@ it needs no external dependencies.
 
 `imganalysis` reads images in PNG format.
 
-    ./imganalysis in example_inputs/GRF_matern_C5.png out data.txt
+    ./imganalysis in example_inputs/GRF_matern_C5.png out outdata.txt
 
 See `./imganalysis help` and <https://morphometry.org/software/papaya2/> for further information.
 
@@ -33,7 +33,7 @@ See `./ppanalysis help` and <https://morphometry.org/software/papaya2/> for furt
 `banana` analyzes astrophysics data in FITS format.  The `CCfits` library is required to build,
 see the end of this document.  After successful compilation with `make banana` it can be used like
 
-    ./banana in example_inputs/SIE_detail2.fits out SIE_detail2_out.dat mint 0.0003 maxt 0.003 numt 100
+    ./banana in example_inputs/SIE_detail2.fits out SIE_detail2_out.txt mint 0.0003 maxt 0.003 numt 100
 
 See `./banana help` and <https://morphometry.org/software/papaya2/> for further information.
 
@@ -44,7 +44,7 @@ See `./banana help` and <https://morphometry.org/software/papaya2/> for further 
 and computes Minkowski Tensors of their excursion sets.
 
     make sersic
-    ./sersic scan_angle threshold 1.5 aspect 0.3 resolution 300 interpolated_marching_squares >sersic0.3.txt
+    ./sersic scan_angle threshold 1.5 aspect 0.3 resolution 300 interpolated_marching_squares >sersic0.3out.txt
 
 The `sersic` example does not require external dependencies.
 
