+++ demos/banana.cpp
@@ -21,11 +21,6 @@ struct FitsFile : Photo
         // FIXME extract metadata from FITS to convert pixel coordinates
         // to sky coordinates.  for now, we map the entire image
         // to the unit square.
-        // FIXME readAllKeys often throws a warning of the type
-        //       Fits Error: Keyword not found: HISTORY
-        //       Fits Error: Keyword not found: COMMENT
-        // at this point, which is harmless.  find a way to disable it.
-        phdu.readAllKeys();
         set_coordinates(0, 0, 1, 1, phdu.axis(0), phdu.axis(1));
         std::valarray<double> buffer;
         phdu.read(buffer);
