+++ matlab/README.md
@@ -1,5 +1,7 @@
 # Matlab examples
 
+These extensions were tested with Matlab R2018b.  They require the data API introduced in this version.
+
 See `imganalysis.m` for an example script that builds and uses the image analysis Matlab extension.
 
 See `ppanalysis.m` for an example script that builds and uses the point pattern analysis Matlab extension.
