### 2020-09-17 b8ab4b7
add note about primary HDU
