### 2020-09-17 6a774f0
minor cleanups
