### 2020-09-26 7c3650f
remove Werror from demo Makefile
