### 2020-09-26 193c3de
remove Werror from demo Makefile
