### 2020-09-26 78cd0e0
add documentation about adding include and lib paths
