### 2020-10-06 #2538 @rmjarvis
I like the changes to the paper.  The new examples gave me a much better feel for the potential uses of the software.  And I'm satisfied with the updates to the documentation and examples.
On my system, `make banana` still had trouble with CCfits, but that's probably some kind of versioning issue, since I used conda rather than brew -- I'm happy to assume that the recommended brew installation would work fine.
So I'm pleased to recommend this for publication.
