### 2020-10-07 0ea2edf
A few minor fixes
