### 2020-10-07 ad1f3b3
remove d3-voronoi acknowledgment since that is Morphometer-only
