### 2020-10-07 #2538 @skapfer
We're excited to see this move forward :partying_face:

I've archived the current state of things (as of the end of the review process) aka git tag 2020-10-07 on Zenodo with DOI

    10.5281/zenodo.4071544

It now lives here: https://zenodo.org/record/4071544

We're thinking of 'Irreducible Minkowski Tensors' (IMTs for short) as a proper name, which is why we titlecased it. If that is not possible for some reason, we prefer titlecasing the whole title, i.e. "papaya2: 2D Irreducible Minkowski Tensor Computation".

I've checked the author list and title on Zenodo agree with the paper.

Cheers,

Sebastian
