### 2020-10-08 #2538 @hugoledoux
> We're thinking of 'Irreducible Minkowski Tensors' (IMTs for short) as a proper name, which is why we titlecased it. If that is not possible for some reason, we prefer titlecasing the whole title, i.e. "papaya2: 2D Irreducible Minkowski Tensor Computation".

If IMT is in the paper also then of course it's fine. Now "irreducible" is not capitalised in the paper! I'm good with anything, as long as consistent.
