### 2020-10-08 #2538 @skapfer
@hugoledoux Good point. I've fixed capitalization the paper!
