### 2020-10-12 #2538 @hugoledoux
@skapfer one last thing needs to be done, you have a release but no version <vX.Y.Z>, since there is none so far I propose naming the release v1.0.0?

I guess it means updating the zenodo too, so please send me back the DOI if modified. Once this is done, we will formally accept it and your paper should be publish within a few days.
