### 2020-10-12 #2538 @skapfer
Good point, the publication is a good point to start semantic versioning. I've tagged v1.0.0 and updated the Zenodo record. The DOI didn't change.
