### 2020-10-12 #2538 @hugoledoux
@skapfer the editor-in-chief will now process/finalise the paper. And it should be published pretty soon.
