+++ paper/paper.bib
@@ -1,6 +1,3 @@
-@STRING{AAP = {Astronomy \& Astrophysics}}
-
-
 @article{bib:Mantz2008,
   author={Hubert Mantz and Karin Jacobs and Klaus Mecke},
   title={Utilizing {Minkowski} functionals for image analysis: a marching square algorithm},
@@ -107,7 +104,7 @@
   author = {M. Kerscher and {Mecke, K.} and {Schmalzing, J.} and {Beisbart, C.} and {Buchert, T.} and {Wagner, H.}},
   title = {Morphological fluctuations of large-scale structure: The {PSCz} survey},
   DOI= "10.1051/0004-6361:20010604",
-  journal = {A\&A},
+  journal = {{Astronomy \& Astrophysics}},
   year = 2001,
   volume = 373,
   number = 1,
